+++ requirements_versions.txt
@@ -5,7 +5,7 @@ basicsr==1.4.2
 blendmodes==2022
 clean-fid==0.1.35
 einops==0.4.1
-fastapi==0.94.0
+fastapi>=0.100.0
 gfpgan==1.3.8
 gradio==3.32.0
 httpcore==0.15
